Clean up custom DG run scripts:
- move calculate_num_order to run_dg_conv_study.py to reduce dependencies in run_dg_utils.py
- use time.process_time instead of deprecated time.clock
